+++ src/console_gui.cpp
@@ -17,6 +17,7 @@
 #include "string_func.h"
 #include "strings_func.h"
 #include "gfx_func.h"
+#include "gfx_layout.h"
 #include "settings_type.h"
 #include "console_func.h"
 #include "rev.h"
