+++ src/gfx_layout.h
@@ -185,4 +185,7 @@ public:
 	static void ReduceLineCache();
 };
 
+Point GetCharPosInString(std::string_view str, const char *ch, FontSize start_fontsize = FS_NORMAL);
+ptrdiff_t GetCharAtPosition(std::string_view str, int x, FontSize start_fontsize = FS_NORMAL);
+
 #endif /* GFX_LAYOUT_H */
