+++ src/textbuf.cpp
@@ -14,6 +14,7 @@
 #include "strings_func.h"
 #include "gfx_type.h"
 #include "gfx_func.h"
+#include "gfx_layout.h"
 #include "window_func.h"
 #include "core/alloc_func.hpp"
